added checks and errors for account errors
and deleted tests

+++ src/main/java/com/ostiasolutions/performancetestapplication/services/AccountService.java
@@ -3,7 +3,9 @@ package com.ostiasolutions.performancetestapplication.services;
 import com.ostiasolutions.performancetestapplication.entities.Account;
 import com.ostiasolutions.performancetestapplication.repositories.AccountRepository;
 import jakarta.transaction.Transactional;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
@@ -17,18 +19,15 @@ public class AccountService {
     }
 
     public Account getAccount(long accountNumber) {
-
-        Optional<Account> account = accountRepository.findDistinctByAccountNumber(accountNumber);
-
-        return account.orElse(null);
+        return accountRepository.findDistinctByAccountNumber(accountNumber)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found"));
     }
 
     public Account createAccount(Account account) {
-
-        Optional<Account> existingAccount = accountRepository.findDistinctByFirstNameIgnoreCaseAndLastNameIgnoreCaseAndAccountNumber(account.getFirstName(), account.getLastName(), account.getAccountNumber());
+        Optional<Account> existingAccount = accountRepository.findDistinctByAccountNumber(account.getAccountNumber());
 
         if (existingAccount.isPresent()) {
-            return null;
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Account already exists with this name and account number.");
         }
 
         return accountRepository.save(account);
@@ -36,17 +35,22 @@ public class AccountService {
 
     @Transactional
     public void deleteAccount(long accountNumber) {
+        Optional<Account> optionalAccount = accountRepository.findDistinctByAccountNumber(accountNumber);
+        if (optionalAccount.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found");
+        }
+
         accountRepository.deleteByAccountNumber(accountNumber);
     }
 
     public Account updateAccount(Account account) {
-        Optional<Account> existingAccount = accountRepository.findDistinctByFirstNameIgnoreCaseAndLastNameIgnoreCaseAndAccountNumber(account.getFirstName(), account.getLastName(), account.getAccountNumber());
+        Optional<Account> existingAccount = accountRepository.findDistinctByAccountNumber(account.getAccountNumber());
 
         if (existingAccount.isPresent()) {
             account.setId(existingAccount.get().getId());
             return accountRepository.save(account);
         } else {
-            return null;
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found");
         }
     }
 }
